Added a give quarry block command
A command that gives a player an iron block named "Quarry".
Added a filter to check that the placed block is the correct quarry block.
Moved the "checkForChest" method in to the run method of the quarry so a player can place a chest next to the quarry at any point