quarry upgrades
New command "/mcquarry giveupgrade <player> <upgrade>". This will give an upgrade item which can be placed inside the new quarry upgrade menu. Current known issue if that there is an error if the player tries to put nothing in to menu and the upgrade item is dropped when used instead of removed