Fillers
Added a new type of quarry, the Filler. This will take blocks from a chest and place them in a selected area. The give command has changed to "/mcquarry give  <player> <type> <level>". Type can be "quarry" or "filler". This is currently not finished. Things that need to be finished are supporting doublechests, filtering the blocks taken from the chest, different levels / speeds and maybe upgrades, but probably not