Filler levels and files
The fillers have 3 different levels (the same as the quarry) and they are saved to a file so they still work when a server restarts. Also added a check to make sure an item in the chest is a block, if not it will not place it. Maybe in the next version the filler could pause if it does not detect anything in the chest or the chest is destroyed